Add new services and refactor project structure
