+++ Argon.Server.sln
@@ -29,6 +29,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Grains.Interface", "src\Gra
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Models", "src\Models\Models.csproj", "{9D390396-E330-4060-9C2A-FD5B933C7F90}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Grains.Impl", "src\Grains.Impl\Grains.Impl.csproj", "{ADC6379C-C252-4C24-9047-E09D7BE73D5E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services", "src\Services\Services.csproj", "{8BEE378B-511C-49C6-AE76-BC04517FA808}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -59,6 +63,14 @@ Global
 		{9D390396-E330-4060-9C2A-FD5B933C7F90}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9D390396-E330-4060-9C2A-FD5B933C7F90}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9D390396-E330-4060-9C2A-FD5B933C7F90}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ADC6379C-C252-4C24-9047-E09D7BE73D5E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ADC6379C-C252-4C24-9047-E09D7BE73D5E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ADC6379C-C252-4C24-9047-E09D7BE73D5E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ADC6379C-C252-4C24-9047-E09D7BE73D5E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8BEE378B-511C-49C6-AE76-BC04517FA808}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8BEE378B-511C-49C6-AE76-BC04517FA808}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8BEE378B-511C-49C6-AE76-BC04517FA808}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8BEE378B-511C-49C6-AE76-BC04517FA808}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -70,6 +82,8 @@ Global
 		{151002BD-57E5-440B-9CA7-C681A69CA02C} = {43C63C3C-90EB-4AED-BB15-B30F76F93A9E}
 		{5BA1F3B1-D442-4C16-B8D5-D9A1D2D13957} = {43C63C3C-90EB-4AED-BB15-B30F76F93A9E}
 		{9D390396-E330-4060-9C2A-FD5B933C7F90} = {43C63C3C-90EB-4AED-BB15-B30F76F93A9E}
+		{ADC6379C-C252-4C24-9047-E09D7BE73D5E} = {43C63C3C-90EB-4AED-BB15-B30F76F93A9E}
+		{8BEE378B-511C-49C6-AE76-BC04517FA808} = {43C63C3C-90EB-4AED-BB15-B30F76F93A9E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {576333C3-629F-47F8-BE21-3E82EDC2655D}
