+++ src/Argon.Api/Argon.Api.csproj
@@ -45,8 +45,10 @@
 
     <ItemGroup>
         <ProjectReference Include="..\Argon.Contracts\Argon.Contracts.csproj"/>
+        <ProjectReference Include="..\Grains.Impl\Grains.Impl.csproj"/>
         <ProjectReference Include="..\Grains.Interface\Grains.Interface.csproj"/>
         <ProjectReference Include="..\ServiceDefaults\ServiceDefaults.csproj"/>
+        <ProjectReference Include="..\Services\Services.csproj" />
     </ItemGroup>
 
     <ItemGroup>
@@ -68,8 +70,12 @@
     </ItemGroup>
 
     <ItemGroup>
+        <Folder Include="Entities\"/>
         <Folder Include="Features\Otp\"/>
         <Folder Include="Grains.Interfaces\"/>
+        <Folder Include="Grains.Persistence.States\"/>
+        <Folder Include="Grains\"/>
+        <Folder Include="Services\"/>
     </ItemGroup>
 
 </Project>
