+++ src/Argon.Api/Features/Jwt/JwtOptions.cs
@@ -1,26 +1,8 @@
 namespace Argon.Api.Features.Jwt;
 
+using Contracts;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.IdentityModel.Tokens;
-
-public record JwtOptions
-{
-    public required string Issuer { get; set; }
-
-    public required string Audience { get; set; }
-
-    // TODO use cert in production
-    public required string   Key     { get; set; }
-    public required TimeSpan Expires { get; set; }
-
-    public void Deconstruct(out string issuer, out string audience, out string key, out TimeSpan expires)
-    {
-        audience = Audience;
-        issuer   = Issuer;
-        key      = Key;
-        expires  = Expires;
-    }
-}
 
 public static class JwtFeature
 {
