+++ src/Argon.Contracts/Argon.Contracts.csproj
@@ -26,4 +26,8 @@
         <PackageReference Include="Microsoft.Orleans.Serialization.Abstractions" Version="8.2.0"/>
     </ItemGroup>
 
+    <ItemGroup>
+        <ProjectReference Include="..\Models\Models.csproj"/>
+    </ItemGroup>
+
 </Project>
