+++ src/Grains.Impl/EmailManager.cs
@@ -2,9 +2,11 @@ namespace Argon.Api.Grains;
 
 using System.Net;
 using System.Net.Mail;
-using Features.EmailForms;
 using global::Grains.Interface;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Orleans;
+using Services;
 
 public class EmailManager(IOptions<SmtpConfig> smtpOptions, ILogger<EmailManager> logger, EMailFormStorage formStorage) : Grain, IEmailManager
 {
