+++ src/Grains.Impl/FusionGrain.cs
@@ -2,6 +2,8 @@
 
 using System.IdentityModel.Tokens.Jwt;
 using Microsoft.IdentityModel.Tokens;
+using Orleans;
+using Orleans.Runtime;
 using Persistence.States;
 
 public class FusionGrain(
