+++ src/Grains.Impl/ServerManager.cs
@@ -5,6 +5,7 @@ using global::Grains.Interface;
 using Microsoft.EntityFrameworkCore;
 using Models;
 using Models.DTO;
+using Orleans;
 
 public class ServerManager(IGrainFactory grainFactory, ApplicationDbContext context) : Grain, IServerManager
 {
