+++ src/Grains.Impl/SessionManager.cs
@@ -2,12 +2,12 @@ namespace Argon.Api.Grains;
 
 using Contracts.etc;
 using Entities;
-using Features.Otp;
 using global::Grains.Interface;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 using Models;
 using Models.DTO;
-using Services;
+using Orleans;
 
 public class SessionManager(
     IGrainFactory grainFactory,
