+++ src/Models/DTO/FlagNameAttribute.cs
@@ -1,0 +1,7 @@
+namespace Argon.Api.Features.Sfu;
+
+[AttributeUsage(AttributeTargets.Field)]
+public class FlagNameAttribute(string flagName) : Attribute
+{
+    public string FlagName { get; } = flagName;
+}
