+++ src/Services/IPasswordHashingService.cs
@@ -1,5 +1,7 @@
 namespace Argon.Api.Services;
 
+using System.Security.Cryptography;
+using System.Text;
 using Features.Otp;
 using Models;
