+++ src/Services/UserInteractionService.cs
